fix: ajust serializer field

+++ bank/serializers/transaction_serializer.py
@@ -7,7 +7,7 @@ from bank.serializers.account_serializer import AccountSerializer
 class TransactionSerializer(serializers.Serializer):
     forma_pagamento = serializers.ChoiceField(required=True, choices=TRANSACTION_TYPES)
     conta_id = serializers.IntegerField(required=True)
-    valor = serializers.IntegerField(required=True)
+    valor = serializers.FloatField(required=True)
 
 
 class TransactionModelSerializer(serializers.ModelSerializer):
